refactor: split lists handler into submodules
